Porting video resnet
