+++ torchvision/prototype/models/__init__.py
@@ -1,2 +1,1 @@
 from . import segmentation
-from . import video
