--- torchvision/prototype/models/video/__init__.py
@@ -1,1 +1,0 @@
-from .resnet import *
